pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blue-bench"
version = "0.1.0"
description = "AI-augmented Blue Team validation framework + MCP tool server for self-hosted open-weight LLMs"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Max Kovalsky"}]
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "ollama>=0.3",
    "anthropic>=0.96",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
blue-bench = "blue_bench_cli.main:app"

[tool.hatch.build.targets.wheel]
packages = [
    "blue_bench_mcp",
    "blue_bench_client",
    "blue_bench_cli",
    "blue_bench_eval",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
